Reorganize to isolate dependencies.
