+++ bluesky_browser/frameworks/qt/summary.py
